Bottom thread also changed to #10-32 UNF
Both top and bottom threads are now the same (with changes to the bracket to follow). The undercut/gutter was made smaller (all test runs so far were producing something far too large). The threaded bottom part is now shorter by 0.7mm (so the bracket can be 100% compatible with the original IBM version).