Enlarged AM5 threads; underside IBM spec compliant
Since directly fitting an AM5 cooler needs a heat spreader due to the hight difference, the threads were purposefully enlarged to break direct compatibility (and require the use of the mounting posts). And since it's no longer restricted to the #10-32 UNC threads a stronger UNF was chosen, allowing a shorter depth and full compliance with IBM's footprint specification.

Note: a known bug in BricsCAD broke the associated measurements ruining most of the 2D drawing. These will be fixed at a later date.