Radiator inlet flow added and dimensions tightened
A lot of tests and attempts to get here. The reservoir hanger was 3D printed and found to be little too tight, so everything moved closer to the fans by 1mm (enough to give the screws clearance). The routing for the hard tubing is correct and mostly free of the airflow.