Present the SCA 0.3mm lower
Testing in the Octane shows the SCA plug to be slightly too high, which was then compared with how a real disk presents then adjusted. Other tweaks were made to mimic how a disk attaches to the sled (by not having completely flat sides), and to avoid bulging the print after cutting the UNC threads.